python-api/requirements.txt
# Core API
fastapi
uvicorn[standard]
sqlalchemy
alembic
python-dotenv
psycopg2-binary

# Auth
python-jose[cryptography]
python-multipart
email-validator

# Data processing
pandas
numpy
openpyxl
requests
qrcode[pil]
boto3

# Scheduling
apscheduler

# AI/ML
ultralytics
hsemotion-onnx
face-recognition
google-generativeai
pdfplumber
mediapipe
opencv-python-headless
